Tease apart the custom SYB from ExactPrint
