+++ ghcide/ghcide.cabal
@@ -140,6 +140,7 @@ library
         include
     exposed-modules:
         Control.Concurrent.Strict
+        Generics.SYB.GHC
         Development.IDE
         Development.IDE.Main
         Development.IDE.Core.Actions
